Ensure consistent dtypes for time params.

+++ gkp/gkp_tf_env/oscillator_env.py
@@ -45,9 +45,9 @@ class OscillatorGKP(Oscillator, BatchOperatorMixinBCH, GKP):
             N (int, optional): Size of oscillator Hilbert space. Defaults to 100.
         """
         self._N = N
-        self.t_round = tf.constant(t_gate + t_read)
-        self.t_feedback = tf.constant(t_feedback)
-        self.t_idle = tf.constant(t_idle)
+        self.t_round = tf.constant(t_gate + t_read, dtype=tf.float32)
+        self.t_feedback = tf.constant(t_feedback, dtype=tf.float32)
+        self.t_idle = tf.constant(t_idle, dtype=tf.float32)
         self.step_duration = tf.constant(t_gate + t_read + t_feedback + t_idle)
         super().__init__(*args, N=N, **kwargs)
